+++ src/irvine/oeis/a117/A117266.java
@@ -17,6 +17,6 @@ public class A117266 extends LambdaSequence {
   /** Construct the sequence. */
   public A117266() {
     super(0, n -> (n <= 1) ? Z.valueOf(n + 1) : ((n == 2) ? Z.SIX : Rationals.SINGLETON.sum(0, n, k -> new Q(Functions.FACTORIAL.z(2 * (n - k)),
-      Functions.FACTORIAL.z(n - k).square()).multiply(Z.ONE.shiftLeft((long) (n - k) * (n + k - 3) / 2))).num()));
+      Functions.FACTORIAL.z(n - k).square()).multiply(Z.ONE.shiftLeft((n - k) * (n + k - 3) / 2))).num()));
   }
 }
