+++ src/irvine/oeis/a367/A367236.java
@@ -17,7 +17,7 @@ public class A367236 extends LambdaSequence {
       final long s = 2;
       final long t = 0;
       final long u = 2;
-      return Integers.SINGLETON.sum(0, n, k -> Binomial.binomial((long) t * k + (long) u * (n - k) + 1, k).multiply(Binomial.binomial(n + (long) (s - 1) * k - 1, n - k)).divide((long) t * k + (long) u * (n - k) + 1));
+      return Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(t * k + u * (n - k) + 1, k).multiply(Binomial.binomial(n + (s - 1) * k - 1, n - k)).divide(t * k + u * (n - k) + 1));
     });
   }
 }
