+++ src/irvine/oeis/a367/A367281.java
@@ -17,7 +17,7 @@ public class A367281 extends LambdaSequence {
       final long s = 3;
       final long t = 2;
       final long u = 3;
-      return Integers.SINGLETON.sum(0, n, k -> Binomial.binomial((long) t * k + (long) u * (n - k) + 1, k).multiply(Binomial.binomial(n + (long) (s - 1) * k - 1, n - k)).divide((long) t * k + (long) u * (n - k) + 1));
+      return Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(t * k + u * (n - k) + 1, k).multiply(Binomial.binomial(n + (s - 1) * k - 1, n - k)).divide(t * k + u * (n - k) + 1));
     });
   }
 }
