+++ src/irvine/oeis/a367/A367284.java
@@ -17,7 +17,7 @@ public class A367284 extends LambdaSequence {
       final long s = 3;
       final long t = 1;
       final long u = 3;
-      return Integers.SINGLETON.sum(0, n, k -> Binomial.binomial((long) t * k + (long) u * (n - k) + 1, k).multiply(Binomial.binomial((long) s * k, n - k)).divide((long) t * k + (long) u * (n - k) + 1));
+      return Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(t * k + u * (n - k) + 1, k).multiply(Binomial.binomial(s * k, n - k)).divide(t * k + u * (n - k) + 1));
     });
   }
 }
