+++ src/irvine/oeis/a368/A368079.java
@@ -17,7 +17,7 @@ public class A368079 extends LambdaSequence {
       final long s = 2;
       final long t = 3;
       final long u = 3;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial((long) t * (n + 1) + k - 1, k).multiply(Binomial.binomial((long) (u + 1) * (n + 1) - (long) s * k - 2, n - (long) s * k))).divide(n + 1);
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * (n + 1) + k - 1, k).multiply(Binomial.binomial((u + 1) * (n + 1) - s * k - 2, n - s * k))).divide(n + 1);
     });
   }
 }
