+++ src/irvine/oeis/a369/A369477.java
@@ -17,7 +17,7 @@ public class A369477 extends LambdaSequence {
       final long s = 2;
       final long t = 2;
       final long u = 1;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial((long) t * (n + 1), k).multiply(Binomial.binomial((long) (t + u) * (n + 1) - k, n - (long) s * k))).divide(n + 1);
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * (n + 1), k).multiply(Binomial.binomial((t + u) * (n + 1) - k, n - s * k))).divide(n + 1);
     });
   }
 }
