+++ src/irvine/oeis/a370/A370159.java
@@ -17,7 +17,7 @@ public class A370159 extends LambdaSequence {
       final long s = 2;
       final long t = 2;
       final long u = 1;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial((long) t * n, k).multiply(Binomial.binomial((long) (t + u) * n - k, n - (long) s * k)));
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * n, k).multiply(Binomial.binomial((t + u) * n - k, n - s * k)));
     });
   }
 }
