+++ src/irvine/oeis/a370/A370215.java
@@ -17,7 +17,7 @@ public class A370215 extends LambdaSequence {
       final long s = 3;
       final long t = 2;
       final long u = 2;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial((long) t * n + k - 1, k).multiply(Binomial.binomial((long) u * n, n - (long) s * k)));
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * n + k - 1, k).multiply(Binomial.binomial(u * n, n - s * k)));
     });
   }
 }
