+++ src/irvine/oeis/a372/A372476.java
@@ -17,7 +17,7 @@ public class A372476 extends LambdaSequence {
       final long s = 3;
       final long t = 1;
       final long u = 1;
-      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial((long) t * n + k - 1, k).multiply(Binomial.binomial((long) (t + u + 1) * n - (long) (s - 1) * k - 1, n - (long) s * k)));
+      return Integers.SINGLETON.sum(0, n / s, k -> Binomial.binomial(t * n + k - 1, k).multiply(Binomial.binomial((t + u + 1) * n - (s - 1) * k - 1, n - s * k)));
     });
   }
 }
