+++ src/irvine/oeis/a378/A378890.java
@@ -19,7 +19,7 @@ public class A378890 extends LambdaSequence {
       final long s = -1;
       final long t = 5;
       final long u = 2;
-      return Rationals.SINGLETON.sum(0, n, k -> new Q(Binomial.binomial((long) t * k + (long) u * (n - k) + r, k)).multiply(Binomial.binomial((long) s * k, n - k)).divide((long) t * k + (long) u * (n - k) + r)).multiply(r).num();
+      return Rationals.SINGLETON.sum(0, n, k -> new Q(Binomial.binomial(t * k + u * (n - k) + r, k)).multiply(Binomial.binomial(s * k, n - k)).divide(t * k + u * (n - k) + r)).multiply(r).num();
     });
   }
 }
